+++ src/main/java/com/github/games647/securemyaccount/commands/EnableCommand.java
@@ -1,14 +1,10 @@
 package com.github.games647.securemyaccount.commands;
 
+import com.github.games647.securemyaccount.Account;
 import com.github.games647.securemyaccount.ImageDownloader;
 import com.github.games647.securemyaccount.SecureMyAccount;
 import com.github.games647.securemyaccount.TOTP;
-import com.google.common.base.Charsets;
-import com.google.common.io.Files;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.logging.Level;
@@ -20,11 +16,11 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class CreateKeyCommand implements CommandExecutor {
+public class EnableCommand implements CommandExecutor {
 
     private final SecureMyAccount plugin;
 
-    public CreateKeyCommand(SecureMyAccount plugin) {
+    public EnableCommand(SecureMyAccount plugin) {
         this.plugin = plugin;
     }
 
@@ -42,14 +38,16 @@ public class CreateKeyCommand implements CommandExecutor {
     }
 
     private void generateKey(Player player) {
-        File keyStorage = new File(plugin.getDataFolder(), player.getUniqueId().toString());
-        if (keyStorage.exists()) {
+        Account account = plugin.getOrLoadAccount(player);
+        if (account.isRegistered()) {
             player.sendMessage(ChatColor.DARK_RED + "You have already a secret key");
             return;
         }
 
         String secretKey = TOTP.generateSecretKey();
-        if (!saveKey(keyStorage, secretKey)) {
+        account.setSecretCode(secretKey);
+        account.setIp(player.getAddress().getHostString());
+        if (!plugin.saveAccount(player)) {
             player.sendMessage(ChatColor.DARK_RED + "Error while saving your secret key");
             return;
         }
@@ -69,29 +67,5 @@ public class CreateKeyCommand implements CommandExecutor {
             plugin.getLogger().log(Level.SEVERE, "Tried creating barcode url", ex);
             player.sendMessage(ChatColor.DARK_RED + "Error occurred while creating the image url");
         }
-    }
-
-    private boolean saveKey(File file, String secretKey) {
-        BufferedWriter fileWriter = null;
-        try {
-            file.createNewFile();
-
-            fileWriter = Files.newWriter(file, Charsets.UTF_8);
-            fileWriter.write(secretKey);
-            fileWriter.flush();
-            return true;
-        } catch (IOException ex) {
-            plugin.getLogger().log(Level.SEVERE, "Tried saving your secret key", ex);
-        } finally {
-            if (fileWriter != null) {
-                try {
-                    fileWriter.close();
-                } catch (IOException ex) {
-                    plugin.getLogger().log(Level.SEVERE, "Tried closing file", ex);
-                }
-            }
-        }
-
-        return false;
     }
 }