+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Added config
 * Added server ip config property
+* Added ip auto login
+* Added protected commands property
+* Added protection for other events besides commands
+* Added configurable commandOnly protection
 * Added /register alias for a new account
 * Added /createkey alias for a new account